🔧 Rename `auth` to `application`
