+++ impressive_strawberry/web/deps/__init__.py
@@ -2,10 +2,10 @@
 This package contains all usable dependencies of :mod:`impressive_strawberry`.
 """
 
-from .auth import *
+from .application import *
 from .database import *
 
 __all__ = (
-    "auth",
+    "application",
     "database",
 )
